Added the new orders button which utilized the NewOrder method in my CustomerOrdersData.cs file to insert a new entry into the newly created audit table and the sales.orders and sales.order_items tables for a given customer.
In the method, the SQL string being run was an execution of a Procedure in SSMS that used a trigger to add an entry to the audit table whenever an entry was added to or updated on the sales.orders table.  This procedure was meant to test that the trigger functioned as intended.

The new orders button, in addition to adding the order to the database, also redirects to the CustomerOrders page for that customer so you can see their new orders page if it falls within the date range specified inside of the GetCustomerOrders method in the CustomerOrdersData.cs file.

I made sure that everything worked when all parameters were filled with legitimate data and that error messages were returned on the page when invalid data was entered.

Lastly, I added the queries for the audit table creation, the trigger to add an entry to the audit table whenever the sales.orders table was updated or a new entry was inserted into it, and the procedure I created to test my trigger into my Text_file_with_Header_and_Repository file.

Also deleted an unused using line in the CustomerOrdersData.cs file to differentiate it from the code that was committed and pushed for the master branch.